Run different animations when loading a scene vs reloading
Took a bit of thinkering to get 2 differents to play one at a time.
However, my braine remember that this is a Finite State Machine.

Created a new init state for ending and starting, and added a bool in the transitions.
I put a static variable in the scene loader so we can start animations correctly.
Works as expected!